Update `index.js`
This commit adds the new feature that's been suggested in the feedback from previous commits. The server will now end the response automatically if the after the specified timeout in the configuraition object the outgoing message [response] is still not ended. As in the feedback and the comment for that block of code, the appropriate status for this response is yet to be discussed. The status "408 Request Timeout" is technically speaking not appropriate in this context, but statuses that would directly apply to this problem are seemingly devoid as well.