Update `config.yml`
As per the feedback from the last commit of this file, having the `blank_issues_enabled` property set to `false` isn't particularly convenient for repositories created suing this template repository when no issue templates are configured. So, this commit will change this pro tempore, until issue templates using the new YAML syntax (as per the GitHub docs) are set up.