Update `index.js`
This commit brings no additions/deletetions of functionality, rather adding reformatting/refactoring of the current source code and changing the documentation comments around. This state of the server is regarded to be ready for a beta pre-release and no significant changes are to be made (aside from bug and security vulnerabilities fixes, of course).