Update `index.js`
This commit adds a small yet huge fix, which is the support for the client (presumbaly being a Webpack application) to surf without page reloads. The change just added a default "root" (HTML) file at the end of the array of possible path queries on the file system, so if no match for a file has been found, it will always send the "default.html" file from the client.