Update `index.js`
This commit adds a little feature for clients to be able to display any 404 Not Found errors at all. From now on, whenever the client requests the root index file from the client, an X-File-Type header will be sent with its value set to "root". This way, if the client make a requests through the Fetch API and gets the root index file, it now knows that it requested a path that does not exist on the server file system in actuality.